+++ src/main.rs
@@ -94,6 +94,13 @@ fn main() {
                 history::add(&mut sh, &mut rl, &line, status, tsb, tse);
             }
             Ok(ReadResult::Eof) => {
+                if let Ok(x) = env::var("NO_EXIT_ON_CTRL_D") {
+                    log!("NO_EXIT_ON_CTRL_D: {:?}", x);
+                    if x == "1" {
+                        println!("");
+                        continue;
+                    }
+                }
                 println!("exit");
                 break;
             }